Implemented reading parameters from config.yaml and adding them to ui, and when changed writing them to config.yaml

+++ IGVC_UI/parameter.cpp
@@ -1,4 +1,4 @@
 #include "parameter.h"
 
-parameter::parameter(const QString& label, int min, int max, int initial)
+parameter::parameter(const QString& label, double min, double max, double initial)
     : label(label), min(min), max(max), initial(initial) {}
